Added button to cancel an armed sensor-based grab.
Cancel either an armed state or a grabbed state by forcing SetBallGrabber(false) to run. This will interrupt the GrabBallWhenSensed command if it is running, whether it is in either sWAITING or sCLOSED. If the grabber was closed, this has the effect of opening it. Pressing the manual grab button would also do this, but if the we were in sWAITING, you'd get a glitch as the grabber tries to close while the button is down. This setup avoids that behavior.